+++ embedded-sdmmc/CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog] and this project adheres to [Semantic 
 
 ### Changed
 
+- Moved a lot of fundamental SD card traits and types into a separate `embedded-sdmmc-types` crate.
+  The library and its `blockdevice` module which mostly retain the same structure are re-exported.
 - Updated to `defmt` 1.0.1, `embedded-hal-bus` 0.3.0, `env_logger` 0.11.8, `heapless` 0.9.1, and `hex-literal` 1.0.0.
 - Raised the minimum supported Rust version to 1.87.0.
 - Removed `core-error` feature as MSRV is now above 1.81